[REFACTOR] web-client: extract task card component
