+++ client/web-client/src/pages/tasks-page.tsx
@@ -1,28 +1,10 @@
-import {
-    Card,
-    CardContent,
-    CardDescription,
-    CardHeader,
-    CardTitle,
-} from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-
-import { Trash2Icon, CheckCircleIcon, EditIcon } from "lucide-react";
-import { format } from "date-fns";
+import { CheckCircleIcon, ClipboardListIcon } from "lucide-react";
 import AddTaskDialog from "@/components/tasks/add-task-dialog";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Badge } from "@/components/ui/badge";
 import { axiosInstance } from "@/lib/utils";
 import { useQuery } from "@tanstack/react-query";
-
-interface Task {
-    _id: string;
-    title: string;
-    description: string;
-    priority: "low" | "medium" | "high";
-    status: "todo" | "in-progress" | "completed";
-    dueDate: Date;
-}
+import type { Task } from "@/lib/types";
+import TaskCard from "@/components/tasks/task-card";
 
 export default function TasksPage() {
     const fetchTasks = async (): Promise<Task[]> => {
@@ -35,12 +17,6 @@ export default function TasksPage() {
         queryKey: ["tasks"],
     });
 
-    const priorityColors = {
-        high: "text-red-500 bg-red-100",
-        medium: "text-yellow-500 bg-yellow-100",
-        low: "text-green-600 bg-green-100",
-    };
-
     return (
         <div className="p-10">
             <div className="flex items-center justify-between mb-8">
@@ -50,70 +26,21 @@ export default function TasksPage() {
 
             <Tabs defaultValue="todo">
                 <TabsList className="w-[200px] bg-blue-50 border border-blue-100 mb-6">
-                    <TabsTrigger value="todo">To Do</TabsTrigger>
-                    <TabsTrigger value="completed">Done</TabsTrigger>
+                    <TabsTrigger value="todo">
+                        <ClipboardListIcon className="w-4 h-4 text-blue-500" />
+                        <span>To Do</span>
+                    </TabsTrigger>
+                    <TabsTrigger value="completed">
+                        <CheckCircleIcon className="w-4 h-4 text-green-500" />
+                        <span>Done</span>
+                    </TabsTrigger>
                 </TabsList>
                 <TabsContent value="todo">
                     <div className="grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4 gap-5">
                         {tasks
                             ?.filter((task) => task.status === "todo")
                             .map((task) => (
-                                <Card key={task._id} className="mb-4">
-                                    <CardHeader className="flex flex-row items-start justify-between pb-2">
-                                        <div className="flex items-center">
-                                            <CardTitle className="text-base font-medium">
-                                                {task.title}
-                                            </CardTitle>
-                                        </div>
-                                        <Badge
-                                            variant={"outline"}
-                                            className={
-                                                priorityColors[task.priority]
-                                            }
-                                        >
-                                            {task.priority}
-                                        </Badge>
-                                    </CardHeader>
-                                    <CardContent className="space-y-2">
-                                        <CardDescription className="text-sm text-gray-600 dark:text-gray-400">
-                                            {task.description}
-                                        </CardDescription>
-                                        <div className="flex items-center justify-between">
-                                            <span className="text-xs text-gray-500">
-                                                {format(
-                                                    task.dueDate,
-                                                    "MMM dd, yyyy"
-                                                )}
-                                            </span>
-                                        </div>
-                                        <div className="flex space-x-2 mt-4">
-                                            <Button
-                                                size="sm"
-                                                // onClick={() =>
-                                                //     updateTaskStatus(
-                                                //         task._id,
-                                                //         "completed"
-                                                //     )
-                                                // }
-                                                className="flex-1 text-white bg-blue-600 hover:bg-blue-500 transition-colors duration-200"
-                                            >
-                                                Mark As Done
-                                            </Button>
-                                            <Button size="sm" variant="outline">
-                                                <EditIcon className="w-4 h-4" />
-                                            </Button>
-                                            <Button
-                                                size="sm"
-                                                variant="outline"
-                                                // onClick={() =>
-                                                //     deleteTask(task._id)
-                                                // }
-                                            >
-                                                <Trash2Icon className="w-4 h-4" />
-                                            </Button>
-                                        </div>
-                                    </CardContent>
-                                </Card>
+                                <TaskCard key={task._id} task={task} />
                             ))}
                     </div>
                 </TabsContent>
@@ -122,66 +49,7 @@ export default function TasksPage() {
                         {tasks
                             ?.filter((task) => task.status === "completed")
                             .map((task) => (
-                                <Card
-                                    key={task._id}
-                                    className="mb-4 border-green-200 dark:border-green-800 bg-green-50 dark:bg-green-900/20"
-                                >
-                                    <CardHeader className="flex flex-row items-start justify-between pb-2">
-                                        <div className="flex items-center">
-                                            <CardTitle className="text-base font-medium line-through text-gray-600 dark:text-gray-400 flex items-center">
-                                                {task.title}
-                                                <CheckCircleIcon className="w-4 h-4 ml-2 text-green-500" />
-                                            </CardTitle>
-                                        </div>
-                                        <Badge
-                                            variant={"outline"}
-                                            className={
-                                                priorityColors[task.priority]
-                                            }
-                                        >
-                                            {task.priority}
-                                        </Badge>
-                                    </CardHeader>
-                                    <CardContent className="space-y-2">
-                                        <CardDescription className="text-sm text-gray-600 dark:text-gray-400 line-through">
-                                            {task.description}
-                                        </CardDescription>
-                                        <div className="flex items-center justify-between">
-                                            <span className="text-xs text-gray-500">
-                                                {format(
-                                                    task.dueDate,
-                                                    "MMM dd, yyyy"
-                                                )}
-                                            </span>
-                                        </div>
-                                        <div className="flex space-x-2 mt-4">
-                                            <Button
-                                                size="sm"
-                                                // onClick={() =>
-                                                //     updateTaskStatus(
-                                                //         task._id,
-                                                //         "todo"
-                                                //     )
-                                                // }
-                                                className="flex-1 text-white bg-blue-600 hover:bg-blue-500 transition-colors duration-200"
-                                            >
-                                                Mark As To Do
-                                            </Button>
-                                            <Button size="sm" variant="outline">
-                                                <EditIcon className="w-4 h-4" />
-                                            </Button>
-                                            <Button
-                                                size="sm"
-                                                variant="outline"
-                                                // onClick={() =>
-                                                //     deleteTask(task._id)
-                                                // }
-                                            >
-                                                <Trash2Icon className="w-4 h-4" />
-                                            </Button>
-                                        </div>
-                                    </CardContent>
-                                </Card>
+                                <TaskCard key={task._id} task={task} />
                             ))}
                     </div>
                 </TabsContent>
